% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compression.R
\name{compress}
\alias{compress}
\title{Compresses a segregation matrix}
\usage{
compress(data, neighbors, group, unit, weight = NULL, max_iter = Inf)
}
\arguments{
\item{data}{A data frame.}

\item{neighbors}{A data frame with exactly two columns. Each row identifies
a set of "neighbors" that may be merged.}

\item{group}{A categorical variable
contained in \code{data}. Defines the first dimension
over which segregation is computed.}

\item{unit}{A categorical variable
contained in \code{data}. Defines the second dimension
over which segregation is computed.}

\item{weight}{Numeric. Only frequency weights are allowed.
(Default \code{NULL})}

\item{max_iter}{Maximum number of iterations (Default \code{Inf})}
}
\value{
Returns a data.table.
}
\description{
Given a data set that identifies suitable neighbors for merging,
this function will merge units iteratively, where in each iteration
the neighbors with the smallest effect on total M will be merged.
}
