% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compression.R
\name{compress_crosswalk}
\alias{compress_crosswalk}
\title{Takes the results from \code{compress} and returns a crosswalk}
\usage{
compress_crosswalk(iterations, n_iterations = Inf)
}
\arguments{
\item{iterations}{A data frame returned by the \code{compress} function.}

\item{n_iterations}{The number of iterations to take into account when
constructing the crosswalk.'}
}
\value{
Returns a data.table.
}
\description{
Given the results from \code{compress}, this function can be used
to get a crosswalk to identify merged units at a given iteration.
}
