% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compression.R
\name{get_crosswalk}
\alias{get_crosswalk}
\title{Create crosswalk after compression}
\usage{
get_crosswalk(compression, n_units = NULL, percent = NULL, parts = FALSE)
}
\arguments{
\item{compression}{A "segcompression" object returned by \link{compress}.}

\item{n_units}{Determines the number of merges by specifying the number of
units to remain in the compressed dataset.
Only \code{n_units} or \code{percent} must be given. (default: \code{NULL})}

\item{percent}{Determines the number of merges by specifying the percentage
of total segregation information retained in the compressed dataset.
Only \code{n_units} or \code{percent} must be given. (default: \code{NULL})}

\item{parts}{(default: FALSE)}
}
\value{
Returns a ggplot2 plot.

Returns a data.table.
}
\description{
After running \link{compress}, this function creates a crosswalk table.
Usually it is preferred to call \link{merge_units} directly.
}
