% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\name{isolation}
\alias{isolation}
\title{Calculates isolation}
\usage{
isolation(data, group, unit, weight = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{group}{A categorical variable
contained in \code{data}. Defines the first dimension
over which segregation is computed.}

\item{unit}{A vector of variables
contained in \code{data}. Defines the second dimension
over which segregation is computed.}

\item{weight}{Numeric. (Default \code{NULL})}
}
\value{
Returns a data.table with group column and isolation index.
}
\description{
Returns isolation index of each group
}
