% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutual.R
\docType{package}
\name{mutual}
\alias{mutual}
\alias{mutual-package}
\title{mutual: Calculation of the Mutual Information Index}
\description{
Computes the mutual information index (M), a measure
of segregation that is highly decomposable. Provides
tools to decompose the index by units and groups, and by within
and between terms. Includes standard error estimation by bootstrapping.
}
\section{Methods}{


\itemize{
 \item{}{\code{\link{mutual_total}} Computes M.}
 \item{}{\code{\link{mutual_local}} Computes local segregation.}
}
}

\section{Data}{


\itemize{
 \item{}{\code{\link{usschools}} Example dataset.}
}
}

