% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutual_diff.R
\name{mutual_diff_mrc}
\alias{mutual_diff_mrc}
\title{Decomposes the difference between two M indices}
\usage{
mutual_diff_mrc(data1, data2, unit, group, weight = NULL, se = FALSE,
  n_bootstrap = 10)
}
\arguments{
\item{data1}{A data frame with same structure as \code{data2}.}

\item{data2}{A data frame with same structure as \code{data1}.}

\item{unit}{A categorical variable or a vector of variables
contained in \code{data}. Defines the first dimension
over which segregation is computed.}

\item{group}{A categorical variable or a vector of variables
contained in \code{data}. Defines the second dimension
over which segregation is computed.}

\item{weight}{Numeric. Only frequency weights are allowed.
(Default \code{NULL})}

\item{se}{If \code{TRUE}, standard errors are estimated via bootstrap.
(Default \code{FALSE})}

\item{n_bootstrap}{Number of bootstrap iterations. (Default \code{10})}
}
\value{
Returns a list with the following elements:
  \code{M1} contains the M for \code{data1}.
  \code{M2} contains the M for \code{data2}.
  \code{diff} is the difference between \code{M2} and \code{M1}.
  \code{unit_entropy} is the difference in unit entropy.
  \code{group_dist} is the contribution of group composition differences.
  \code{invariant} is the contribution of group composition-invariant differences.
  If \code{se} is \code{TRUE}, each element will be a vector of length two, with
  the second item containing the associated bootstrapped standard error.
}
\description{
Uses the method developed by Mora and Ruiz-Castillo (2009)
}
\references{
Ricardo Mora and Javier Ruiz-Castillo. 2009. "The Invariance Properties of the
  Mutual Information Index of Multigroup Segregation". Research on Economic Inequality 17: 33-53.
}
