% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutual_difference.R
\name{mutual_difference}
\alias{mutual_difference}
\title{Decomposes the difference between two M indices}
\usage{
mutual_difference(data1, data2, unit, group, weight = NULL, method = NULL,
  se = FALSE, n_bootstrap = 50)
}
\arguments{
\item{data1}{A data frame with same structure as \code{data2}.}

\item{data2}{A data frame with same structure as \code{data1}.}

\item{unit}{A categorical variable or a vector of variables
contained in \code{data}. Defines the first dimension
over which segregation is computed.}

\item{group}{A categorical variable or a vector of variables
contained in \code{data}. Defines the second dimension
over which segregation is computed.}

\item{weight}{Numeric. Only frequency weights are allowed.
(Default \code{NULL})}

\item{method}{Either "mrc" (Mora and Ruiz-Castillo, 2009) or "elbers" (Elbers, 2018)}

\item{se}{If \code{TRUE}, standard errors are estimated via bootstrap.
(Default \code{FALSE})}

\item{n_bootstrap}{Number of bootstrap iterations. (Default \code{50})}
}
\value{
Returns a data frame with columns \code{stat} and \code{est}. The data frame contains
  the following rows defined by \code{stat}:
  \code{M1} contains the M for \code{data1}.
  \code{M2} contains the M for \code{data2}.
  \code{diff} is the difference between \code{M2} and \code{M1}.

  If \code{method} is "mrc", the data frame contains the following additional rows:
  \code{unit_entropy} is the difference in unit entropy.
  \code{group_marginal} is the contribution of group composition differences.
  \code{invariant} is the contribution of group composition-invariant differences.
  Note that \code{diff = unit_entropy + group_marginal + invariant}

  If \code{method} is "elbers", the data frame contains the following additional rows:
  \code{group_marginal} is the contribution of the change in group marginals.
  \code{unit_marginal} is the contribution of the change in unit marginals.
  \code{conditional} is the contribution of the change in the conditional probabilities.
  \code{additions} is the weighted local linkage for \code{group} elements not in \code{data1}.
  \code{removals} is the negative weighted local linkage for \code{group} elements not in \code{data2}.
  Note that \code{diff = group_marginal + unit_marginal + conditional + additions + removals}

  If \code{se} is set to \code{TRUE}, an additional column \code{se} contains
  the associated bootstrapped standard errors, and the column \code{est} contains
  bootstrapped estimates.
}
\description{
Uses either the method developed by Mora and Ruiz-Castillo (2009) or
the method developed by Elbers (2018)
}
\examples{
mutual_difference(usschools00, usschools05, unit="race", group="school",
    weight="n", method="mrc")
mutual_difference(usschools00, usschools05, unit="race", group="school",
    weight="n", method="elbers")
}
\references{
Ricardo Mora and Javier Ruiz-Castillo. 2009. "The Invariance Properties of the
  Mutual Information Index of Multigroup Segregation". Research on Economic Inequality 17: 33-53.

Elbers, Benjamin. 2018. An Alternative Difference Decomposition of M. Working Paper.
}
