% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutual.R
\name{mutual_local}
\alias{mutual_local}
\title{Calculates local segregation indices}
\usage{
mutual_local(data, unit, group, weight = NULL, se = FALSE,
  n_bootstrap = 10)
}
\arguments{
\item{data}{A data frame.}

\item{unit}{A categorical variable or a vector of variables
contained in \code{data}. Defines the dimension
over which segregation is computed.}

\item{group}{A categorical variable or a vector of variables
contained in \code{data}. Defines the group for which local
segregation indices are calculated.}

\item{weight}{Numeric. Only frequency weights are allowed.
(Default \code{NULL})}

\item{se}{If \code{TRUE}, standard errors are estimated via bootstrap.
(Default \code{FALSE})}

\item{n_bootstrap}{Number of bootstrap iterations. (Default \code{10})}
}
\value{
Returns a data frame, with one row for each of the categories
  defined by \code{group}. The first column(s) define the \code{group}.
  The column \code{ls} contains the local segregation index, with a
  column \code{ls_se} containing the standard error (if \code{se} is \code{TRUE}).
  The column \code{p} contains the proportion of each group from the total
  number of cases. The column \code{M_group} calculates \code{ls * p}. Standard
  errors for this column are contained in \code{M_group_se} if requested.
}
\description{
Returns local segregation indices for each category defined
by \code{group}.
}
\examples{
# which racial groups are most segregated?
(localseg = mutual_local(usschools, 'school', 'race', weight='n'))
# native americans are most segregated, whites are least segregated.

sum(localseg[, 'p']) # => 1

# the sum of the weighted local segregation scores equals
# total segregation
mutual_total(usschools, 'school', 'race', weight='n') # => .422
sum(localseg[, 'M_group']) # => .422
}
\references{
Henri Theil. 1971. Principles of Econometrics. New York: Wiley.

Ricardo Mora and Javier Ruiz-Castillo. 2011. "Entropy-based Segregation Indices". Sociological Methodology 41(1): 159–194.
}
