% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{school_ses}
\alias{school_ses}
\title{Student-level data including SES status}
\format{
A data frame with 5,153 rows and 3 variables:
\describe{
  \item{school_id}{school ID}
  \item{ethnic_group}{one of A, B, or C}
  \item{ses_quintile}{SES of the student (1 = lowest, 5 = highest)}
}
}
\usage{
school_ses
}
\description{
Fake dataset used for examples. This is an individual-level
dataset of students in schools.
}
\keyword{datasets}
