% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{schools05}
\alias{schools05}
\title{Ethnic/racial composition of schools for 2005/2006}
\format{A data frame with 8,013 rows and 5 variables:
\describe{
  \item{state}{either A, B, or C}
  \item{district}{school agency/district ID}
  \item{school}{school ID}
  \item{race}{either native, asian, hispanic, black, or white}
  \item{n}{n of students by school and race}
}}
\usage{
schools05
}
\description{
Fake dataset used for examples. Loosely based on data provided by
the National Center for Education Statistics, Common Core of Data,
with information on U.S. primary schools in three U.S. states.
The original data can be downloaded at \url{https://nces.ed.gov/ccd/}.
}
\keyword{datasets}
