% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compression.R
\name{scree_plot}
\alias{scree_plot}
\title{Scree plot for segregation compression}
\usage{
scree_plot(compression, tail = Inf)
}
\arguments{
\item{compression}{A "segcompression" object returned by \link{compress}.}

\item{tail}{Return only the last \code{tail} units (default: \code{Inf})}
}
\value{
Returns a ggplot2 plot.
}
\description{
A plot that allows to visually see the effect of compression
on mutual information.
}
