% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{segcurve}
\alias{segcurve}
\title{A visual representation of two-group segregation}
\usage{
segcurve(data, group, unit, weight)
}
\arguments{
\item{data}{A data frame.}

\item{group}{A categorical variable or a vector of variables
contained in \code{data}. Defines the first dimension
over which segregation is computed.}

\item{unit}{A categorical variable or a vector of variables
contained in \code{data}. Defines the second dimension
over which segregation is computed.}

\item{weight}{Numeric. (Default \code{NULL})}
}
\value{
Returns a ggplot2 object.
}
\description{
Produces a segregation curve, as defined in Duncan and Duncan (1955)
}
