% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_plot.R
\name{seg_plot}
\alias{seg_plot}
\title{A visual representation of segregation}
\usage{
seg_plot(data, group, unit, weight, order = "segregation")
}
\arguments{
\item{data}{A data frame.}

\item{group}{A categorical variable or a vector of variables
contained in \code{data}. Defines the first dimension
over which segregation is computed.}

\item{unit}{A categorical variable or a vector of variables
contained in \code{data}. Defines the second dimension
over which segregation is computed.}

\item{weight}{Numeric. (Default \code{NULL})}

\item{order}{A character, either "segregation" or "majority". Affects
the ordering of the units.}
}
\value{
Returns a ggplot2 object.
}
\description{
Produces a segregation plot.
}
