% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segregation.R
\docType{package}
\name{segregation}
\alias{segregation}
\alias{segregation-package}
\title{segregation: Entropy-based segregation indices}
\description{
Computes entropy-based segregation indices, with a focus on
the mutual information index (M). The M is a measure
of segregation that is highly decomposable. Provides
tools to decompose the index by units and groups, and by within
and between terms. Includes standard error estimation by bootstrapping.
}
\section{Methods}{


\itemize{
 \item{}{\code{\link{mutual_total}} Computes M.}
 \item{}{\code{\link{mutual_local}} Computes local segregation based on M.}
 \item{}{\code{\link{mutual_difference}} Decomposes difference between two M indices.}
}
}

\section{Data}{


\itemize{
 \item{}{\code{\link{schools00}} and \code{\link{schools05}} Example datasets.}
}
}

