% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segregation.R
\docType{package}
\name{segregation}
\alias{segregation}
\alias{segregation-package}
\title{segregation: Entropy-based segregation indices}
\description{
Calculate and decompose entropy-based, multigroup segregation indices, with a focus
on the Mutual Information Index (M) and Theil's Information Index (H).
Provides tools to decompose the measures by groups and units, and by within
and between terms. Includes standard error estimation by bootstrapping.
}
\seealso{
\url{https://elbersb.com/segregation}
}
\author{
\strong{Maintainer}: Benjamin Elbers \email{be2239@columbia.edu} (\href{https://orcid.org/0000-0001-5392-3448}{ORCID})

}
